refactor(config): rename components/universal_validators.py to field_coercion.py
Clean break — old file deleted, all imports updated.
